Remove debug prints and optimize imports
