+++ shell/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter/services.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:shell/display/display.widget.dart';
 import 'package:shell/display/monitor/screen/screen.provider.dart';
@@ -20,8 +19,6 @@ void main() {
   SchedulerBinding.instance.addPostFrameCallback((_) {
     //platformApi.startupComplete();
   });
-
-  RawKeyboard.instance.addListener(print);
 
   container
     ..read(waylandManagerProvider)
